Fix errno: 150 "Foreign key constraint is incorrectly formed"

+++ database/migrations/2022_10_14_010501_add_purchase_order_date_warranty_date_purchase_cost_column_to_serials_table.php
@@ -18,7 +18,7 @@ class AddPurchaseOrderDateWarrantyDatePurchaseCostColumnToSerialsTable extends M
             $table->date('purchase_date')->nullable();
             $table->date('warranty_date')->nullable();
             $table->decimal('purchase_cost', 10, 2)->nullable();
-            $table->bigInteger('supplier_id')->unsigned()->nullable();
+            $table->unsignedInteger('supplier_id')->unsigned()->nullable();
 
             $table->foreign('supplier_id')->references('id')->on('suppliers');
         });
